requirements.txt
fastapi
uvicorn
openai
pinecone
sentence-transformers
httpx
beautifulsoup4
SQLAlchemy
python-dotenv
